+++ pkg/commands/objectstoreentry/doc.go
@@ -1,0 +1,3 @@
+// Package objectstoreentry contains commands to inspect and manipulate Fastly edge
+// object stores keys.
+package objectstoreentry
