--- pkg/commands/objectstorekeys/doc.go
@@ -1,3 +1,0 @@
-// Package objectstorekeys contains commands to inspect and manipulate Fastly edge
-// object stores keys.
-package objectstorekeys
